Refactor + add the meal to the PR
