+++ src/main/java/com/coding/restaurant/restaurant/controllers/NewMealController.java
@@ -38,25 +38,7 @@ public class NewMealController implements Initializable {
 
 
   public void addMeal(String name, String description, Double price, String image, Boolean isActive, String type) {
-    try (Connection connexion = ConnectDatabaseController.getConnection();
-         PreparedStatement statement = connexion.prepareStatement("INSERT INTO Meal (UUID, name, description, price, image, isActive,Type) VALUES (?,?,?, ?, ?, ?, ?)")) {
-      statement.setString(1, generateUUID());
-      statement.setString(2, name);
-      statement.setString(3, description);
-      statement.setDouble(4, price);
-      statement.setString(5, image);
-      statement.setBoolean(6, isActive);
-      statement.setString(7, type);
-      statement.executeUpdate();
-      System.out.println("Meal added");
-    } catch (SQLException e) {
-      e.printStackTrace();
-    }
-  }
-
-  //  function generate UUID
-  public static String generateUUID() {
-    return java.util.UUID.randomUUID().toString();
+    DatabaseManager.addMeal(name, description, price, image, isActive, type);
   }
 
   @Override
